pyproject.toml
[project]
name = "ingest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "frikanalen-django-api-client",
    "httpx>=0.28.1",
    "inotify>=0.2.10",
    "pathvalidate>=3.3.1",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "anyio>=4.9.0",
    "black>=25.1.0",
    "click>=8.2.1",
    "datamodel-code-generator>=0.31.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-httpserver>=1.1.3",
    "ruff>=0.12.3",
    "tuspy>=1.1.0",
    "watchdog>=6.0.0",
]

[tool.ruff]
line-length = 120
# generated code files
exclude = ["app/tus_hook/hook_schema.py", "app/ffprobe_schema.py"]
[tool.ruff.lint]
extend-select = [
    "I",     # isort: import sorting
    "F",     # pyflakes: undefined names, unused imports
    "E",     # pycodestyle: basic style issues
    "W",     # pycodestyle: warnings
    "C90",   # mccabe: complexity
    "UP",    # pyupgrade: suggest modern syntax
    "B",     # flake8-bugbear: likely bugs, bad patterns
    "A",     # flake8-builtins: avoid shadowing built-ins
    "SIM",   # flake8-simplify: simplify redundant code
    "TID",   # flake8-tidy-imports: disallow relative imports
    "RUF",   # Ruff-specific rules (e.g. `RUF100` for `__all__` sorted)
]
[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"

[tool.uv.sources]
frikanalen-django-api-client = { path = "../frikanalen-django-api-client" }
